requirements.txt
aiosignal==1.3.1
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.1
cramjam==2.6.2
distlib==0.3.6
fastparquet==2023.2.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
jsonschema==4.17.3
locket==1.0.0
msgpack==1.0.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
partd==1.3.0
platformdirs==3.0.0
proto-plus==1.22.2
protobuf==4.22.0
psutil==5.9.4
py4j==0.10.9.5
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.19.3
pyspark==3.3.2
python-can==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
ray==2.2.0
requests==2.28.2
rsa==4.9
six==1.16.0
toolz==0.12.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.19.0
wrapt==1.14.1
